+++ frontend/components/project/ProjectCarousel.tsx
@@ -2,7 +2,7 @@ import Slider from 'react-slick';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
 import ProjectCard from './ProjectCard';
-import { Project } from '@/types/types';
+import { Project } from '@/types/project';
 
 type Props = {
   projects: Project[];
