+++ frontend/styles/App.css
@@ -87,7 +87,7 @@ a {
 
 .main-btn {
   font-size: medium;
-  border-radius: 4px;
+  border-radius: var(--radius-def);
   background: #256ce1;
   padding: 12px 22px;
   color: #fff;
