+++ frontend/types/types.ts
@@ -3,31 +3,3 @@
 export type DefaultObj = {
   [key: string]: string;
 };
-
-export type job = {
-  [key: string]: { want: number; current: number };
-};
-
-export type Project = {
-  id: number;
-  author: string;
-  start: Date;
-  end: Date;
-  tags: string[];
-  stacks: string[];
-  jobs: job[];
-  title: string;
-  content: string;
-  state: 1 | 2 | 3 | 4;
-  createAt: string;
-  view: number;
-  heart: number;
-  comment: never[];
-  position?: string;
-};
-
-export type PostState = {
-  id: number;
-  heart: boolean;
-  want: string;
-};
